requirements.txt
# Fundamentals
numpy
tqdm
scipy
pytorch
audio2numpy
transformers
librosa